Added basic TileGrid
Implemented the Basic TileGrid Objects: TileGrid, Tile, Tilebased. Implemented a debugging Class called: Redsquare, which could later be transformed into our hoveranimation or whatever we like. It currently works as displaying that the basic logic works and that the tiles are there and can be hovered, even if they do not have sprites yet, because our sprite system is not yet implemented