Saving and loading
Created fully functional save and load system, with auto saves and simple usability. All gameobjects in GameLoop.gameObjects are periodically saved and are immediately loaded in properly once we start up the game. I had some code that automatically filled the map with tiles, but now that code is unnecessary for now, because once they were created, they were saved to the SavaData Text file and are now immediately loaded in everytime we open the game. This will allow us to manipulate the map in real time, when drawing our tiles on the map and it will allow us to save the progress of a player, when he closes and reopens the game