Added update method functionality experimental zoom
Everything written in the update method of a gameObject will happen every frame. Scrolling with mousewheel makes very inaccureate zooming happen with tons of rounding errors. closing the program will completely reset all of this. As far as I am aware there is no implementation that possibly makes this go away. So it is just expermiental for developing the map. Not to be used for the player.