More comments, bug fixes, healthbars and health
Documentation has been added for everything written so far. Fixed tons of bugs relating to persistance, object placement, stability and performance. Healthbars and health have been implemented for structures and Characters(which we still have none of). To test the feature you need to have a structure selected and try to place it on top of another structure. This will deal one damage to the existing structure and display its health bar. With consecutive click the healthbar will go down, until it reaches zero, when the structure will be deleted completely.