Improved movement, added very simple object avoidance
Do not trust my object avoidance algorithm. It is the best I could do. I will not improve it, as I have wasted way too much time on that shit. It will work 80% of the time if we have non structure tiles that are impassible for our boy. So be scarce on things he is not supposed to go through.