Added costs to structures and infinite waves
We are this close to have a working game. Now we just need to save the wavecount and money, with the GUI that should be a complete product, now that our economy is working. I also fixed tons of concurrent modification exceptions again. Only errors youg get now are at the very start if you press start too quickly. Also added saving and pausing methods, that we can use in GUI development.